# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Veterinario",
    "version": "1.0.0",
    "description": "API de Veterinario que gestiona usuarios y animales"
    
  },
  "servers": [
    {
      "url": "http://localhost:5002",
      "description": "Servidor de desarrollo"
    }
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "summary": "Lista de todos los usuarios",
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Usuario"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
 
 "/animales": {
      "get": {
        "summary": "Lista de todos los animales",
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Animal"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/animales-usuarios": {
      "get": {
        "summary": "Lista de usuarios y animales relacionados",
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AnimalUsuario"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuario-animal": {
      "post": {
        "summary": "Crear un nuevo usuario y un nuevo animal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mail": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "adress": {
                    "type": "string"
                  },
                  "nameAnimal": {
                    "type": "string"
                  },
                  "specie": {
                    "type": "string"
                  },
                  "breed": {
                    "type": "string"
                  }
                },
                "required": ["name", "mail", "nameAnimal", "specie", "breed"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario y animal creados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuario/{id}": {
      "put": {
        "summary": "Actualizar un usuario existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "mail": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "adress": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Recurso actualizado"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Eliminar un usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario eliminado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al eliminar el usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario no eliminado"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          }
        }
      },
      "Animal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "especie": {
            "type": "string"
          },
          "raza": {
            "type": "string"
          },
          "fk_usuario": {
            "type": "integer"
          }
        }
      },
      "AnimalUsuario": {
        "type": "object",
        "properties": {
          "animal_id": {
            "type": "integer"
          },
          "animal_nombre": {
            "type": "string"
          },
          "animal_especie": {
            "type": "string"
          },
          "animal_raza": {
            "type": "string"
          },
          "usuario_id": {
            "type": "integer"
          },
          "usuario_nombre": {
            "type": "string"
          },
          "usuario_email": {
            "type": "string"
          }
        }
      }
    }
  }
}